+++ terraform/variables.tf
@@ -194,6 +194,18 @@ variable "meld_api_url" {
   sensitive   = true
 }
 
+variable "callstatic_api_key" {
+  description = "Callstatic API key"
+  type        = string
+  sensitive   = true
+}
+
+variable "zan_api_key" {
+  description = "Zan API key"
+  type        = string
+  sensitive   = true
+}
+
 #-------------------------------------------------------------------------------
 # Analytics